+++ src/resources/lang/en/langfilemanager.php
@@ -15,4 +15,6 @@ return [
 
 	'rules_text'	=> "<strong>Notice: </strong> Do not translate words prefixed with colon (ex: ':number_of_items'). Those will be replaced automatically with a proper value. If translated, that stops working.",
 
+	'empty_file'    => "No translations available.",
+
 ];